Changing artitecture of app
The last artitecture was not so good. So in this I changed the way Web Sockets are used in Project and. Now the artitecture is Client-Server-Client. The data is send to main Server which would then emit the Data to raspberry PI. The rest main Server and left i would post on Github in Future, Thanks